Move 'SelectedVersion::from_user_input()' to 'install' command
